chore: categorize action creators by type
