+++ src/reducers/teachSessionReducer.ts
@@ -55,8 +55,6 @@ const teachSessionReducer: Reducer<TeachSessionState> = (state = initialState, a
             return { ...initialState, all: state.all.filter(t => t.teachId !== action.teachSessionGUID) }
         case AT.DELETE_MEMORY_FULFILLED:
             return { ...state, memories: [] }
-        case AT.TEACH_MESSAGE_RECEIVED:
-            return { ...state, input: action.message, scoreInput: undefined, scoreResponse: undefined, extractResponses: [] };
         case AT.INIT_MEMORY_FULFILLED:
             return { ...state, memories: action.memories }
         case AT.RUN_EXTRACTOR_FULFILLED:
