+++ src/types/ActionTypes.ts
@@ -135,7 +135,6 @@ export enum AT {
     RUN_SCORER_FULFILLED = 'RUN_SCORER_FULFILLED',
     POST_SCORE_FEEDBACK_ASYNC = 'POST_SCORE_FEEDBACK_ASYNC',
     POST_SCORE_FEEDBACK_FULFILLED = 'POST_SCORE_FEEDBACK_FULFILLED',
-    TEACH_MESSAGE_RECEIVED = 'TEACH_MESSAGE_RECEIVED',
     TOGGLE_AUTO_TEACH = 'TOGGLE_AUTO_TEACH',
 
     SETTINGS_RESET = 'SETTINGS_RESET',
